Added __repr__, __str__, and quick_init functions to the GamePiece class and all subclasses
We can now just read the data from an object by hitting enter. And we can also use the print function.

There is a new class function quick_init which takes in the output of repr(piece) as either a string or a dictionary. It returns a new instance of the class with the correct pieceType.

You can use this function when testing to different intances quickly and easily.

The function also inherits to the subclasses in a smart way